Refactored calculating number of reduce tasks to use to option parser classes.
We had the calculation of the number of reduce tasks to use spread
between option parsing and main application classes.  I tried to make
the behaviour more uniform moving it all to the option parsing classes.
An additional refactoring run should be done to make the app-specific
parsing classes inherit from the SealToolParser---that should avoid much
of the current code redundancy.