Demux: if sample is alone in lane, assign to it all reads from the lane
Up to now demux used to only assign reads to a sample id if the index
sequence matched at least within the specified number of mismatches.
With this commit, if a sample is alone in a lane, demux will assign to
the sample all reads from its lane---regardless of their index sequence.