Apply reflection-based Hadoop 2 API compatibility fix to unit tests
A previous commit had dealt with the TaskAttemptContext change in the
Hadoop API within the TextSampler, which used to try to instantiate it
directly.  This commit refactors that code into a utility function (in
the Utils module) and uses it to fix some unit tests which presented the
same problem.